fix: moving code around
