+++ aidial_adapter_bedrock/llm/model/claude/v3/converters.py
@@ -1,48 +1,19 @@
-import json
-from typing import (
-    Iterable,
-    List,
-    Literal,
-    Optional,
-    Set,
-    Tuple,
-    assert_never,
-    cast,
-)
+from typing import Iterable, List, Literal, Optional, Set, Tuple, assert_never
 
-from aidial_sdk.chat_completion import FinishReason, Tool, ToolCall
-from anthropic.types.beta import (
-    BetaBase64PDFSourceParam as Base64PDFSourceParam,
-)
+from aidial_sdk.chat_completion import FinishReason, Tool
 from anthropic.types.beta import (
     BetaCacheControlEphemeralParam as CacheControlEphemeralParam,
 )
-from anthropic.types.beta import (
-    BetaCitationsConfigParam as CitationsConfigParam,
-)
-from anthropic.types.beta import BetaContentBlock as ContentBlock
 from anthropic.types.beta import BetaContentBlockParam as ContentBlockParam
 from anthropic.types.beta import BetaImageBlockParam as ImageBlockParam
 from anthropic.types.beta import BetaMessageParam as MessageParam
-from anthropic.types.beta import (
-    BetaPlainTextSourceParam as PlainTextSourceParam,
-)
 from anthropic.types.beta import (
     BetaRequestDocumentBlockParam as RequestDocumentBlockParam,
 )
 from anthropic.types.beta import BetaStopReason as ClaudeStopReason
 from anthropic.types.beta import BetaTextBlockParam as TextBlockParam
 from anthropic.types.beta import BetaToolParam as ToolParam
-from anthropic.types.beta import (
-    BetaToolResultBlockParam as ToolResultBlockParam,
-)
-from anthropic.types.beta import BetaToolUseBlockParam as ToolUseBlockParam
 from anthropic.types.beta import BetaUsage as Usage
-from anthropic.types.beta.beta_base64_image_source_param import (
-    BetaBase64ImageSourceParam as Base64ImageSourceParam,
-)
-from pydantic import BaseModel
-from pydantic import ValidationError as PydValidationError
 
 from aidial_adapter_bedrock.dial_api.token_usage import TokenUsage
 from aidial_adapter_bedrock.llm.errors import ValidationError
@@ -55,86 +26,19 @@ from aidial_adapter_bedrock.llm.message import (
     SystemMessage,
 )
 from aidial_adapter_bedrock.llm.model.attachment_processor import (
-    AttachmentProcessor,
     AttachmentProcessors,
 )
+from aidial_adapter_bedrock.llm.model.claude.v3.blocks import (
+    create_text_block,
+    create_tool_result_block,
+    create_tool_use_block,
+)
+from aidial_adapter_bedrock.llm.model.claude.v3.state import (
+    get_message_content_from_state,
+)
 from aidial_adapter_bedrock.llm.tools.tools_config import ToolsMode
-from aidial_adapter_bedrock.utils.concurrency import aiter_to_list
 from aidial_adapter_bedrock.utils.list import group_by
 from aidial_adapter_bedrock.utils.list_projection import ListProjection
-from aidial_adapter_bedrock.utils.log_config import bedrock_logger as log
-from aidial_adapter_bedrock.utils.resource import Resource
-
-_ClaudeFinishReason = Literal[
-    "end_turn", "max_tokens", "stop_sequence", "tool_use"
-]
-
-_ImageMediaType = Literal["image/png", "image/jpeg", "image/gif", "image/webp"]
-
-
-class MessageState(BaseModel):
-    claude_message_content: List[ContentBlock]
-
-    def to_dict(self) -> dict:
-        return self.dict(
-            # FIXME: a hack to exclude the private __json_buf field
-            exclude={"claude_message_content": {"__all__": {"__json_buf"}}},
-            # Excluding `citations: null`, since they could not be even parsed
-            # currently by the Bedrock.
-            exclude_none=True,
-        )
-
-
-def _get_message_content_from_state(
-    idx: int, message: AIRegularMessage | AIToolCallMessage
-) -> List[ContentBlockParam] | None:
-    if (cc := message.custom_content) is not None and (
-        state_dict := cc.state
-    ) is not None:
-        try:
-            state = MessageState.parse_obj(state_dict)
-            return [block.to_dict() for block in state.claude_message_content]  # type: ignore
-        except PydValidationError as e:
-            log.error(
-                f"Invalid state at the path 'messages[{idx}].custom_content.state': {e}"
-            )
-            return None
-
-    return None
-
-
-def _create_text_block(text: str) -> TextBlockParam:
-    return TextBlockParam(text=text, type="text")
-
-
-def _create_image_block(resource: Resource) -> ImageBlockParam:
-    return ImageBlockParam(
-        source=Base64ImageSourceParam(
-            data=resource.data_base64,
-            media_type=cast(_ImageMediaType, resource.type),
-            type="base64",
-        ),
-        type="image",
-    )
-
-
-_CITATIONS_ENABLED = True
-
-
-def _create_text_document_block(
-    resource: Resource,
-) -> RequestDocumentBlockParam:
-    return RequestDocumentBlockParam(
-        source=PlainTextSourceParam(
-            data=resource.data.decode("utf-8"),
-            media_type=resource.type,  # type: ignore
-            type="text",
-        ),
-        type="document",
-        citations=CitationsConfigParam(enabled=_CITATIONS_ENABLED),
-        title="",  # FIXME
-    )
-
 
 _claude_cache_breakpoint = CacheControlEphemeralParam(type="ephemeral")
 
@@ -157,76 +61,21 @@ def _add_cache_control(
     return claude_content
 
 
-def _to_message_param(
+def _get_claude_message_role(
     dial_message: (
         AIRegularMessage
         | AIToolCallMessage
         | HumanRegularMessage
         | HumanToolResultMessage
     ),
-    claude_content: Iterable[ContentBlockParam],
-) -> MessageParam:
+) -> Literal["assistant", "user"]:
     match dial_message:
         case AIRegularMessage() | AIToolCallMessage():
-            role = "assistant"
+            return "assistant"
         case HumanRegularMessage() | HumanToolResultMessage():
-            role = "user"
+            return "user"
         case _:
             assert_never(dial_message)
-    return MessageParam(role=role, content=claude_content)
-
-
-def _create_pdf_document_block(resource: Resource) -> RequestDocumentBlockParam:
-    return RequestDocumentBlockParam(
-        source=Base64PDFSourceParam(
-            data=resource.data_base64,
-            media_type=resource.type,  # type: ignore
-            type="base64",
-        ),
-        type="document",
-        citations=CitationsConfigParam(enabled=_CITATIONS_ENABLED),
-        title="",  # FIXME
-    )
-
-
-IMAGE_ATTACHMENT_PROCESSOR = AttachmentProcessor(
-    supported_types={
-        "image/png": {"png"},
-        "image/jpeg": {"jpeg", "jpg"},
-        "image/gif": {"gif"},
-        "image/webp": {"webp"},
-    },
-    handler=_create_image_block,
-)
-
-PDF_ATTACHMENT_PROCESSOR = AttachmentProcessor(
-    supported_types={"application/pdf": {"pdf"}},
-    handler=_create_pdf_document_block,
-)
-
-TEXT_ATTACHMENT_PROCESSOR = AttachmentProcessor(
-    supported_types={"text/plain": {"txt"}},
-    handler=_create_text_document_block,
-)
-
-
-def _to_claude_tool_call(call: ToolCall) -> ContentBlockParam:
-    return ToolUseBlockParam(
-        id=call.id,
-        name=call.function.name,
-        input=json.loads(call.function.arguments),
-        type="tool_use",
-    )
-
-
-def _to_claude_tool_result(
-    message: HumanToolResultMessage,
-) -> ToolResultBlockParam:
-    return ToolResultBlockParam(
-        tool_use_id=message.id,
-        type="tool_result",
-        content=[_create_text_block(message.content)],
-    )
 
 
 def _merge_messages_with_same_role(
@@ -273,46 +122,39 @@ async def to_claude_messages(
             break
 
         idx_offset += 1
-        content = await aiter_to_list(
-            handlers.process_attachments(_create_text_block, message)
-        )
+        content = await handlers.process_attachments(message)
         content = _add_cache_control(message, content)
         system_messages.extend(content)  # type: ignore
 
-    ret: ListProjection[MessageParam] = ListProjection()
+    claude_messages: ListProjection[MessageParam] = ListProjection()
+
     for idx, message in enumerate(messages[idx_offset:], start=idx_offset):
 
         match message:
             case HumanRegularMessage():
-                content = await aiter_to_list(
-                    handlers.process_attachments(_create_text_block, message)
-                )
+                content = await handlers.process_attachments(message)
 
             case AIRegularMessage():
                 # Take the message content from the state if possible,
                 # since it may include certain content blocks that
                 # are missing from the DIAL message itself,
                 # such as thinking signatures and redacted thinking blocks.
-                content = _get_message_content_from_state(idx, message)
+                content = get_message_content_from_state(idx, message)
                 if content is None:
-                    content = await aiter_to_list(
-                        handlers.process_attachments(
-                            _create_text_block, message
-                        )
-                    )
+                    content = await handlers.process_attachments(message)
 
             case AIToolCallMessage():
-                content = _get_message_content_from_state(idx, message)
+                content = get_message_content_from_state(idx, message)
 
                 if content is None:
                     content = [
-                        _to_claude_tool_call(call) for call in message.calls
+                        create_tool_use_block(call) for call in message.calls
                     ]
-                    if message.content is not None:
-                        content.insert(0, _create_text_block(message.content))
+                    if (text_content := message.content) is not None:
+                        content.insert(0, create_text_block(text_content))
 
             case HumanToolResultMessage():
-                content = [_to_claude_tool_result(message)]
+                content = [create_tool_result_block(message)]
 
             case SystemMessage():
                 raise ValidationError(
@@ -321,12 +163,14 @@ async def to_claude_messages(
             case _:
                 assert_never(message)
 
-        claude_message = _to_message_param(
-            message, _add_cache_control(message, content)
+        claude_message = MessageParam(
+            role=_get_claude_message_role(message),
+            content=_add_cache_control(message, content),
         )
-        ret.append(claude_message, idx)
 
-    return system_messages, _merge_messages_with_same_role(ret)
+        claude_messages.append(claude_message, idx)
+
+    return system_messages, _merge_messages_with_same_role(claude_messages)
 
 
 def to_dial_finish_reason(
@@ -384,16 +228,3 @@ def to_claude_tool_config(tool: Tool) -> ToolParam:
         tool_param["cache_control"] = _claude_cache_breakpoint
 
     return tool_param
-
-
-def get_usage_message(supported_exts: List[str]) -> str:
-    return f"""
-The application answers queries about attached images.
-Attach images and ask questions about them in the same message.
-
-Supported image types: {', '.join(supported_exts)}.
-
-Examples of queries:
-- "Describe this picture" for one image,
-- "What are in these images? Is there any difference between them?" for multiple images.
-""".strip()
